Implemented support for split/partial apks It's implemented similarly to a normal restore using the package manager to install the package. To install a partial package, the base package must be specified. If it's not specified, the method assumes it's a base package.

+++ src/main/java/com/machiav3lli/backup/handler/ShellCommands.java
@@ -91,6 +91,7 @@ public class ShellCommands implements CommandHandler.UnexpectedExceptionListener
         String label = app.getLabel();
         String packageData = app.getDataDir();
         String packageApk = app.getSourceDir();
+        String[] splitApks = app.getSplitSourceDirs();
         String packageName = app.getPackageName();
         String deviceProtectedPackageData = app.getDeviceProtectedDataDir();
         String backupSubDirPath = backupSubDir.getAbsolutePath();
@@ -105,23 +106,20 @@ public class ShellCommands implements CommandHandler.UnexpectedExceptionListener
 
         // Copying APK & DATA
         boolean clearCache = prefs.getBoolean(Constants.PREFS_CLEARCACHE, true);
-        switch (backupMode) {
-            case AppInfo.MODE_APK:
-                commands.add(String.format("cp \"%s\" \"%s\"", packageApk, backupSubDirPath));
-                break;
-            case AppInfo.MODE_DATA:
-                if (clearCache)
-                    commands.add(String.format("%s rm -r /data/data/%s/cache/*", toybox, packageName));
-                commands.add(String.format("cp -r \"%s\" \"%s\"", packageData, backupSubDirPath));
-                break;
-            default: // defaults to MODE_BOTH
-                if (clearCache)
-                    commands.add(String.format("%s rm -r /data/data/%s/cache/*", toybox, packageName));
-                commands.add(String.format("cp -r \"%s\" \"%s\"", packageData, backupSubDirPath));
-                commands.add(String.format("cp \"%s\" \"%s\"", packageApk, backupSubDirPath));
-                break;
+        if(backupMode == AppInfo.MODE_APK || backupMode == AppInfo.MODE_BOTH) {
+            commands.add(String.format("cp \"%s\" \"%s\"", packageApk, backupSubDirPath));
+            if(splitApks != null){
+                Log.i(TAG, label + " is a split apk");
+                for(String splitApk : splitApks){
+                    commands.add(String.format("cp \"%s\" \"%s\"", splitApk, backupSubDirPath));
+                }
+            }
         }
-
+        if(backupMode == AppInfo.MODE_DATA || backupMode == AppInfo.MODE_BOTH) {
+            if (clearCache)
+                commands.add(String.format("%s rm -r /data/data/%s/cache/*", toybox, packageName));
+            commands.add(String.format("cp -r \"%s\" \"%s\"", packageData, backupSubDirPath));
+        }
         // Copying external DATA
         boolean backupExternal = prefs.getBoolean(Constants.PREFS_EXTERNALDATA, true);
         File backupSubDirExternalFiles = null;
@@ -488,14 +486,15 @@ public class ShellCommands implements CommandHandler.UnexpectedExceptionListener
         return 1;
     }
 
-    public int restoreUserApk(File backupDir, String label, String apk, String ownDataDir) {
+    public int restoreUserApk(File backupDir, String label, String apk, String ownDataDir, String basePackage) {
         /* according to a comment in the android 8 source code for
          * /frameworks/base/cmds/pm/src/com/android/commands/pm/Pm.java
          * pm install is now discouraged / deprecated in favor of cmd
          * package install.
          */
-        final String installCmd = Build.VERSION.SDK_INT >= 28 ?
-                "cmd package install" : "pm install";
+        final String installCmd = String.format("%s%s",
+                Build.VERSION.SDK_INT >= 28 ? "cmd package install" : "pm install",
+                basePackage != null ? " -p " + basePackage : "");
         // swapBackupDirPath is not needed with pm install
         List<String> commands = new ArrayList<>();
         /* in newer android versions selinux rules prevent system_server
